Use jagBase::Transform to compute viewProj and modelViewInvTrans matrices in the nodes and jagload demos.

+++ src/tests/texture/texture.cpp
@@ -48,7 +48,7 @@ public:
 
     virtual bool startup( const unsigned int numContexts );
     virtual bool init();
-    virtual bool frame( const gmtl::Matrix44f& view, const gmtl::Matrix44f& proj );
+    virtual bool frame( const gmtl::Matrix44d& view, const gmtl::Matrix44d& proj );
     virtual bool shutdown()
     {
         return( true );
@@ -188,7 +188,7 @@ bool TextureDemo::init()
 }
 
 
-bool TextureDemo::frame( const gmtl::Matrix44f& view, const gmtl::Matrix44f& proj )
+bool TextureDemo::frame( const gmtl::Matrix44d& view, const gmtl::Matrix44d& proj )
 {
     if( !getStartupCalled() )
         return( true );
